+++ banking/snapshot.go
@@ -244,13 +244,14 @@ func (e *Engine) LoadState(ctx context.Context, p *types.Payload) ([]types.State
 func (e *Engine) restoreRecurringTransfers(ctx context.Context, transfers *checkpoint.RecurringTransfers) error {
 	// ignore events here as we don't need to send them
 	_ = e.loadRecurringTransfers(ctx, transfers)
-
+	e.bss.changed[recurringTransfersKey] = true
 	return nil
 }
 
 func (e *Engine) restoreScheduledTransfers(ctx context.Context, transfers []*checkpoint.ScheduledTransferAtTime) error {
 	// ignore events
 	_, err := e.loadScheduledTransfers(ctx, transfers)
+	e.bss.changed[scheduledTransfersKey] = true
 	return err
 }
 
@@ -296,9 +297,7 @@ func (e *Engine) restoreAssetActions(ctx context.Context, aa *types.BankingAsset
 	for _, v := range aa.AssetAction {
 		asset, err := e.assets.Get(v.Asset)
 		if err != nil {
-			e.log.Error("error restoring asset actions for asset", logging.String("asset", v.Asset))
-
-			continue
+			e.log.Panic("trying to restore an assetAction with no asset", logging.String("asset", v.Asset))
 		}
 
 		aa := &assetAction{